+++ examples/agent/CHANGELOG.md
@@ -1,5 +1,19 @@
 # examples-agent
 
+## 1.0.1-alpha.53
+
+### Patch Changes
+
+- Updated dependencies [9066f95]
+  - @mastra/deployer-cloudflare@0.0.1-alpha.11
+
+## 1.0.1-alpha.52
+
+### Patch Changes
+
+- Updated dependencies [0d5a03d]
+  - @mastra/deployer-cloudflare@0.0.1-alpha.10
+
 ## 1.0.1-alpha.51
 
 ### Patch Changes